+++ Wasap/WasapTests/TestAssets.xcassets/OCRTestImage.imageset/OCRTestImage.jpeg
