+++ Wasap/WasapTests/TestAssets.xcassets/OCRTestImage2.imageset/Contents.json
@@ -5,7 +5,7 @@
       "scale" : "1x"
     },
     {
-      "filename" : "testImageOCR.jpeg",
+      "filename" : "OCRTestImage2.jpeg",
       "idiom" : "universal",
       "scale" : "2x"
     },
